rename and reorder

+++ utils/cleaning.sh
@@ -1,51 +1,76 @@
 #!/bin/bash
 
+plot=false
+log=true
 gun=false
-res=false
+drs=false
 hat=false
 tree=false
+calib=false
+recon=false
+anal=false
 highland=false
 all=false
-log=true
 
 while [[ "$#" -gt 0 ]]; do
    case $1 in
+      --nolog) log=false ;;
+      --plot) plot=true ;;
       --gun) gun=true ;;
-      --res) res=true ;;
+      --drs) drs=true ;;
       --hat) hat=true ;;
       --tree) tree=true ;;
+      --calib) calib=true ;;
+      --recon) recon=true ;;
+      --anal) anal=true ;;
       --highland) highland=true ;;
       --all) all=true ;;
-      --nolog) log=false ;;
    esac
    shift
 done
 
 rm -f $HOME/slurm-*.out
-rm -f $HOME/plots/*.pdf
 rm -f $HOME/nd280Geant4Sim.*
 rm -f $HOME/hatRecon/${HATRECONCONFIG}/core*
 
+if [ "$plot" = true ]; then
+   rm -f $HOME/plots/*.pdf
+fi
+
 if [ "$log" = true ]; then
-    rm -f $HOME/public/Output_log/*.log
+   rm -f $HOME/public/output_nd280/logs/*.log
+   rm -f $HOME/public/output_hatRecon/logs/*.log
+   rm -f $HOME/public/output_highland/logs/*.log
 fi
 
 if [ "$gun" = true ] || [ "$all" = true ]; then
-    rm -f $HOME/public/Output_root/*/1_gun_*.root
+   rm -f $HOME/public/data/MC/1_PG_*.root
 fi
 
-if [ "$res" = true ] || [ "$all" = true ]; then
-    rm -f $HOME/public/Output_root/*/2_DetResSim_*.root
+if [ "$drs" = true ] || [ "$all" = true ]; then
+   rm -f $HOME/public/data/MC/*DRS_*.root
 fi
 
 if [ "$hat" = true ] || [ "$all" = true ]; then
-    rm -f $HOME/public/Output_root/*/3_HATRecon_*.root
+   rm -f $HOME/public/output_hatRecon/root/*/3_HATRecon_*.root
+fi
+
+if [ "$calib" = true ] || [ "$all" = true ]; then
+   rm -f $HOME/public/output_nd280/root/*/3_eventCalib_*.root
+fi
+
+if [ "$recon" = true ] || [ "$all" = true ]; then
+   rm -f $HOME/public/output_nd280/root/*/4_eventRecon_*.root
+fi
+
+if [ "$anal" = true ] || [ "$all" = true ]; then
+   rm -f $HOME/public/output_nd280/root/*/*eventAnalysis_*.root
 fi
 
 if [ "$highland" = true ] || [ "$all" = true ]; then
-    rm -f $HOME/public/Output_highland/highland_*.root
+   rm -f $HOME/public/output_highland/highland_*.root
 fi
 
 if [ "$tree" = true ] || [ "$all" = true ]; then
-    rm -f $HOME/public/Output_root/*/*TreeMaker_*.root
+   rm -f $HOME/public/output_hatRecon/root/*TreeMaker_*.root
 fi
